feat(token-watcher): introduce token-watcher service
This commit introduces a new `token-watcher` service that is responsible for determining the correct data source for a given token contract.

The `token-watcher` service exposes an endpoint that returns the data source (either `stxtools` or `stxcity`) for a given token contract. This information is cached in Redis to improve performance.